pyproject.toml
[tool.poetry]
name = "caterpie"
version = "0.1.0"
description = "Lightweight ORM for uploading CSV files to a relational database."
authors = ["smk508 <skhan8@mail.einstein.yu.edu>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.6"
pandas = "^0.25.3"
configparser = "^4.0.2"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"
names = "^0.3.0"
psycopg2-binary = "^2.8.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
